requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.2
psycopg2-binary==2.9.5
gunicorn==20.1.0
Flask-JWT-Extended